Move renderer scripts to class dir
