+++ app/modules/classes/ipc-logger.ts
